refactor(yaml): select a collection builder once, split the loader
The parser produces three shapes — plain objects, `Map`s, or a node tree —
and every reader re-derived which one was in play: eight `state.nodes ?`
branches, four `instanceof` chains, and the same anchor block copied into
all three collection readers.

That cost correctness, not just readability. Each site decided independently
and they could disagree. Merge keys were detected by comparing a key against
the string `"<<"`, which silently stopped matching the moment keys became
`Scalar`s, so `parseDocument` returned an unmerged `<<` key while `parse`
merged. `stringKeys` put a bare string in a `Pair.key`, leaving one un-node
that `visit` skipped and `isScalar` rejected.

`State.build` is now chosen once in the constructor and the readers ask it
for `map()`, `seq()`, `push()`, `set()`, `key()`, `finish()` and `alias()`.
Representation-dependent logic has one home, so the shapes cannot drift
apart again. All three builders are object literals with the same keys in
the same order, so they share a hidden class and the call sites stay
monomorphic.

With the representation decision extracted, three seams in `loader.ts` had
no back-edge into the composer and moved out: `collections.ts` (storing a
pair, merges, source spans), `properties.ts` (tag/anchor/alias tokens) and
`stream.ts` (directives, document markers, the entry points). `loader.ts` is
now the mutually recursive readers and nothing else — 1553 lines to 736.

`collection-builder.ts` imports `State` as a type only; `state.ts` imports
the builder as a value, so a runtime import back would be a cycle.

Conformance unchanged at 402/402 strict and 396/402 loose. Parse throughput
measured neutral: 1.58/1.42/1.33/1.47x js-yaml against a 1.52/1.47/1.34/1.44
baseline across the four workloads.
